docs(readme): mention only the NORC way
twice on different computers I tried the first advised command:

`:source https://raw.githubusercontent.com/nvim-neorocks/rocks.nvim/master/installer.lua`
which prompts:
E484: Can't open file https://raw.githubusercontent.com/nvim-neorocks/rocks.nvim/master/installer.lua

This is a quite diconcerting error message and not easy to troubleshoot imo. In my case seems like it was oil.nvim.

The README is already long so I suggest that instead of having one
command that might work and one command that works we keep the command
that works even if the user has to run a non-customized nvim.
It's easier for him to go from the 2nd command to the first one, and
even in the NORC one, it wont be long, just the time to run the
installer.