fix(installer): use `vim.o.sh` instead of `'sh'` (#179)
On Windows, `sh` might not be available (e.g. if not installed
with git).
Even if git `sh` is installed, it may not see everything that is
on the `PATH` (see #166).
The default `vim.o.sh` on Windows is `cmd.exe`. On Linux and
MacOS, the script should behave as before.